Remap some CommandException errors instead of reporting NORMAL_CLEARING.
When a CommandException occurs during EVENT_GET_LAST_CALL_FAIL_CAUSE, we
currently just report NORMAL_CLEARING.

There are some CommandException error codes which are truly exceptional,
so we will remap these as ERROR_UNSPECIFIED (not the best, but it is
better than NORMAL_CLEARING).  We'll also set the vendorCause to the
commandError code so that it'll appear in the Telecom logs.

Test: Build (not reproducible).
Bug: 114046998
Change-Id: I983d68a264699828355048448022a025d61e771f