Consolidate IMSI Checking+Setting Logic
Move all IMSI range checking and setting logic
to setImsi() so that it's done regardless of the
source of the IMSI. Also consolidate the IMSI-based
MNC length detection logic so that it's common to
both places where it's done. The logic is slightly
reworked so that fewer redundant checks are performed
inside the function - whenever the AD record is loaded
the result will either be a valid MNC or an UNKNOWN.
That way, calls to the MNC detection logic can avoid
checking for UNINITIALIZED. There are also a couple
small null correctness checks done.

Bug: 115398966
Test: atest FrameworksTelephonyTests
Change-Id: I9401026b99a6457bc02edd2a1387781a565ee6fd