Add support for network identified IMS emergency calls.
There is an existing ImsCallProfile property which can be used to identify
that a call is an emergency call (based on what the network reports).
Using this to set a new connection property in Telecom.
Also using a new Telecom API to check if we're in an emergency call when
toggling airplane mode.

Test: Added new test intent to mark ongoing IMS call as network IDed
emergency call.
Test: Used test intent to verify that the new connection property is set
and that the user cannot turn on airplane mode.

Bug: 77565333
Change-Id: I4723eb216d59d339bad697ebb48f337f521ad70b