Expand onCallHandover logging and add data enabled check.
There appears to be a case where a call can handover from WIFI to LTE and
ImsPhoneCallTracker thinks data is disabled, even though it is not.
Best I can tell, the ImsPhoneCallTracker must not have been notified of
the correct data enabled state.  I'm now querying this directly from the
DCTracker when we have a handover scenario and logging if there is a
mismatch.

Test: Manual test.
Bug: 112872341
Bug: 122474550
Change-Id: I9c466fef36965fd8696ae9ace1b979274a4de706
Merged-In: I0a4738fec062ac2d023d1cdd60c2e52ee4d8fcc9