Notify both background call and foreground call about merge failure
When a user initiates a merge of an ims call with an ims conference call
the upper layers (including dialer) is notified about the ongoing merge
for both the foreground call and the background call. However, if the
merge fails, the upper layers will only be notified that the merge has
completed for the background call.

The dialer will hide the On Hold Banner for calls with ongoing merge,
causing the On Hold Banner for the call that was in the foreground to be
forever hidden, if this call is ever put on hold.

Solved by notifying both foreground call and background call that the
merge failed and the merge operation has completed.

Bug: 148260910
Test: atest ImsPhoneCallTrackerTest
Change-Id: I22cd3398e3cfb786a9b6fa607606e706fdda8098