Catch SecurityException in EuiccCardController and trigger callback.
Currently, if checkCallingPackage() throws a SecurityException, it won't
trigger the callback. Since, EuiccCardManager is async which will return
after invoke EuiccCardController, the throwed SecurityException will get
lost. In this CL, the exception is caught in EuiccCardController and
triggers the callback in this case.
Also, fix a typo in EuiccCardManager.

Bug: 78599148
Bug: 110845244
Test: test on phone
Merged-In: I0883d3948793a9312e744277a0374e16d328fd81
Change-Id: I0883d3948793a9312e744277a0374e16d328fd81