Dynamically revoke Internet capability in NetworkAgent.
When we switch Internt data from a subscription where
its Internet and IMS are on the same APN, the dataConnection
is not really torn down because IMS is still needed. Here
we still report capability of Internet + IMS through NetworkAgent.
Therefore the ConnectivityService will stick to that network
and tear down the new one. By revoking the capability
of Internet through NetworkAgent API, this problem is resolved.

Bug: 118347948
Test: manual
Change-Id: I2f8fd0812d1016f87063f645b1ce4ca3bedff276
Merged-In: I2f8fd0812d1016f87063f645b1ce4ca3bedff276