Change PhoneSwitcher registrants to a single instance.
With new PhoneSwitcher design, active status may be
unchanged but a phone can change from default to non-
default or vice versa. In this case, we may fail to
notify the registrants as active status is unchanged.
By changing registrants from a per phone list to single
instance, we make sure all registrants will be notified
whenever there's a change.

Bug: 117609448
Test: unittest
Change-Id: Ia43f329baa8a562649756dfa8ea8b5c9c60a6a8f