PIN state shall never be confused
IccRecords derived classes call UiccCardApplication.registerForLocked()
in order to know whether PIN1 is verified or it is not enabled because
only limited number of EFs like PL, LI and ICCID can be read if PIN1 is
enabled and not verified yet. That means that the registrants shall be
called back only when PIN1 is enabled and not verified yet.

Bug: 118956681
Test: Confirmed the unexpected callback does not happen with this fix.

Change-Id: I5f8edf5209b08cbf88954ede3ef1059c03c48c84