Fix potential crash when updating voicemail phone number.
A recent change was made which put the above call to
AsyncResult.forMessage into an if block cased on a carrier config
setting.  As a result, if the carrier config is not set,
AsyncResult.forMessage will not be called.
This message is received in GsmCdmaPhone and there is an assumption that
that the type will be AsyncResult, however it now ends up being Message.

Test: Code inspection, compile.
Bug: 118538281
Change-Id: If541e497192f4d3e8142495983257e967a30c7a6