Ensure handover from LTE to WIFI event gets sent in some edge cases.
Where a call starts on LTE and WIFI is not available, the user would never
see the "call continued on wifi" toast.
This is due to the fact we suppress the handover message when handover
happens at the start of the call.
To fix this, if there is no wifi available at the start of a call, deem
the "start of call handover" phase to have already ended, and allow the
message to be shown.

Test: Manual
Bug: 65490850
Change-Id: Ic01eac1f2f0aef4229881818f50afe0c14b4e5da