EuiccController/Connector support multiple eUICCs
This is done using cardId, which is then converted into the slotId. The
reason for the conversion is that the EuiccService API was already
defined using the slot index, and we can't change the existing API
without deprectating all the functions.

Bug: 80097562
Test: manual
Change-Id: I6087f26b90960fd26bc989fb71b534d3d92f48e0
Merged-In: I6087f26b90960fd26bc989fb71b534d3d92f48e0