Don't apply Internet network request specified on non-default sub.
Cellular network requests may specify a subscription ID. However
if the requests is asking for Internet capability, network factory
should ignore it if it's not targeted on preferred data phone.

In implementation, TelephonyNetworkFactory is refactored so that
it queries a PhoneSwitcher API on every network request to decide
whether it should be applied or not. This way PhoneSwitcher has
better control of distrubuting network requests more flexibly.

Bug: 121094781
Test: unittest
Change-Id: I49f8702b19c0bd0f559cf9a97b3d0fa41f6f7222
Merged-In: I49f8702b19c0bd0f559cf9a97b3d0fa41f6f7222