Add Emergency Call Routing field
Some countries or carriers require some emergency numbers that must
be handled with normal call routing or emergency routing.

Do not send service types from the database
as the emergency service categoreis

When searching for the corresponding service categories,
try to find the one that comes from the radio indication.

Use the same rule to set Emergency call information for connection and
ImsProfile as the phone selection rule that is applied to
TelephonyConnectionService

Test: Treehugger
Bug: 112657134
Change-Id: Idcaae194cf43ee2c82a20c2359b189a23dc4d387