Fix test so that it runs as a JUnit 4 rather than JUnit 3 test
The test indirectly extended junit.framework.TestCase from JUnit 3 but
used JUnit 4 @Test annotations. That combination caused the runner
to run it as a JUnit 3 test which is almost certainly not what was
wanted. Amongst other issues it means that @Ignore annotations did not
work.

This change stops it extending AndroidTestCase and fixes a problem
with a non test method being annotated with @Test.

Bug: 122886015
Test: atest FrameworksTelephonyTests
Change-Id: I67b334ac898cca05283387832f7a848cc552b33f