UiccController registers for EID loaded
On RadioConfig < 1.2, we don't get the EID with the slot/card status,
which means we can't use it to set mCardStrings or mDefaultEuiccCardId.

In this case, we will postpone the processing for when EID is received
via ADPU commands.

Test: manually verified mDefaultEuiccCardId behavior with DSDS
Bug: 122978614
Change-Id: I70f567bf70a2704e36fd6d8d32b221ea15572828
Merged-In: I70f567bf70a2704e36fd6d8d32b221ea15572828