Restarting the device must be triggered even in the airplane mode
The device does not trigger to restart itself while the radio power is
off in the current implementation, so user can replace SIM cards in the
airplane mode even if the SIM hot swap feature is disabled. The device
must be restarted in that scenario unless it is still in the start-up
procedure.

Bug: 123387537
Test: Manual

Change-Id: I9034cecd1cc7ecb7f6f662bfc02091db5053814c