End the suffering caused by ConnectivityServiceMock.
TelephonyNetworkFactoryTest uses a fake ConnectivityService to
simulate the behaviour of the system adding and removing
NetworkRequests to/from DcTracker.

Unfortunately, ConnectivityServiceMock subclasses the real
ConnectivityService binder interface (IConnectivityManager),
which means that every time we make a change to that interface,
the mock needs to be updated as well or the build will break.
This wastes time and forces unnecessary cross-project changes.

Stop subclassing the binder interface and instead just make
ConnectivityServiceMock a class of its own. At this point its
resemblance to ConnectivityService is very distant anyway.

Also remove the OWNERS file that was required to ensure that the
core networking team could autonomously make changes to the mock.

Test: atest com.android.internal.telephony.dataconnection.TelephonyNetworkFactoryTest
Change-Id: If2ea101c1a69097118dff451fb352ed1d709d4fd