MMTEL Compat: Remove VT flag when VT+VoLTE are enabled
Older modems would not batch feature flags together, so they
would be sent down one at a time. If multiple features are
enabled at one time (VT+VoLTE, for example), it would cause
two registrations on the network. This has been fixed with the
ImsService APIs, but we still need to keep this behavior in
the compat layer for older devices.

Bug: 110130897
Test: Manual, check logs during registration for VT+VoLTE
Change-Id: Ibe17ed4277d8d73f1e5feddb5f12947e134c4677