In CellularNetworkValidator, handle back to back request.
If back to back requests are sent to CellularNetworkValidator,
it will stop the previous validation silently and restart validation
on the latest subscription. To make it work, a per subscription
NetworkCallback will be created for each validation so that when
old validation result comes in, we can compare the validated subId
agains the latest and avoid race condition.

Bug: 122255288
Test: unittest
Change-Id: Ie128e0ab03c4ca3eed55be1a88bda35283abab06