Remove dependency on canonical-addresses
Currently, canonical-addresses is our only dependency on MmsProvider.
Currently our access requirements to the table is also slightly
different than current usage. To avoid modifying permissions to allow
RcsMessageStoreController to reliably access MmsProvider, and to provide
access in the cleanest manner for the RcsMessageStoreController, this
will provide a proxy to the canonical-addresses table specifically for
RCS message store purposes.

Note that the way RCS message storage interacts with canonical-addresses
may change in the future as well.

Test: Tested manually against future CtsRcsTestCases package
Change-Id: I06662d96a8b2979cd11ff05001953c95968278df