Refactor simState & activeSubscriptionInfo
We log a base SimState, ActiveSubscriptionInfo and the
EnabledModemBitmap at the first time and reset. If any of these data is
changed, the change will be logged.

Move ActiveSubscriptionInfo to the TelephonyLog so that all
the telephony event can share these data during the aggregation.

Bug: 123047205
Test: build
Change-Id: I584572c0252d74cc08e39825f79f7477156bdb50
Merged-In: I8e510746b16edd436380fb98a21fbb4fb4faa0b6