Fix the synchronization pattern in IccPhoneBookInterfaceManager.
Earlier it used the same lock on multiple threads, including the
main thread (handleMessage()). That can pontentially lead do the
main thread being stuck and even a deadlock as was the case in
the bug here.

When a request is processed on another thread, the same lock is
acquired on that thread and before it is released some
processing needs to be done on the main thread. But since
that thread holds the lock that main thread needs (to proceed
with handleMessage()), it leads to a deadlock.

Test: basic sanity
Bug: 113663617
Change-Id: I2314f64d81e7d1fe157e69ccb06c94786e0043c1