Stop local ringback on SRVCC.
When a call undergoes SRVCC is may be playing local ringback.  Ensuring
that the local ringback will stop in this case.

Also fixed flakey test in the ImsPhoneCallTest; this drove some changes
in ImsPhoneCall changing how the ImsPhone is retrieved; the
TelephonyTestCase does some black magic to replace instances of things
at test setup.  Realistically there is no reason you can't get just the
phone from the ImsPhoneCallTracker and use that to avoid the black magic.

Test: Added new unit test for this case.
Bug: 119917326
Change-Id: I5e79b410f36d7a56c0a885bd35eecaadf13973db