Fix issue where TTY mode was not being updated to the RIL.
Some overly optimistic code cleanup in another CL resulted in the TTY mode
no longer being updated to the RIL.
The reason was the removal of the TtyManager class in Telephony.
Rather than revert those changes and bring back a bunch of other cruft,
adding the code from TtyManager which is required to GsmCdmaPhone instead.

Test: Manual; verify TTY mode changes in UI are reflected to the modem.
Bug: 126592801
Change-Id: I20ad212eb2b02518fa852c7772c5ca11e45aa569