Adding subscription group owner.
Define owner of a subscription group. Such that when subscriptions are
being added into or removed from a group, as long as callingPackage is
the group owner, it can skip carrier permission check of subscriptions
in the group.

This is necessary when a subscription is being added into a
group that contains inactive subscription, which mean its carrier
privilege can't be verified.

Bug: 131854492
Test: unittest
Change-Id: I67ef976bba43294fab2329219869392868993072
Merged-In: I67ef976bba43294fab2329219869392868993072