Keep old country ISO for country-related emergency number if signal is lost
Sometimes the country is updated as an empty string when the network signal
is lost; though we may not call emergency when there is no signal, we want
to keep the old country iso to provide country-related emergency numbers,
because they think they are still in that country. So we don't need to update
country change in this case.

Bug: 130779927
Test: Sanity
Change-Id: I08f93a5bf20d81d0cf98b87b352b84bb1425ef2b