Make SST.getMdnNumber() thread-safe.
Remove earlier code that was added for debugging. The main
conclusion from that was that mdn can be queried when RUIM is
not provisioned. Incorporated the fix for that in this CL.

Test: atest FrameworksTelephonyTests:ServiceStateTrackerTest#testGetMdn
Bug: 131394428
Merged-in: I490473b1c3d403a67533247c0178813c8f57edb5
Change-Id: I490473b1c3d403a67533247c0178813c8f57edb5
(cherry picked from commit 3d194fd513988282012b31875d7be4938f412ee4)