Make sure when only SIM is inserted in DSDS device, makes it default
1) Fix the bug that when eSIM is not inserted and propriatery pSIM
switches profile, we still ask user to select it as preferred SIM.
The expected behavior is it becomes default sub automatically.

2) Fix the bug that sometimes we don't ask user to re-select data sub.
when both pSIM and eSIM are present and pSIM is default data sub, if
eSIM switches profile we failed to ask user to re-select data sub.

3) Also, defined an enum to define primary sub change for better
readability.

Bug: 132111589
Test: unittest and manual tests
Change-Id: Ifb49dcab601fc74acc6251bd243044f415114a01
Merged-In: Ifb49dcab601fc74acc6251bd243044f415114a01