Update MultiSimSettingControllerTest to use all public APIs
MultiSimSettingController is updated to be handler based. Here
we are updating its unittests to be the same. Otherwise it becomes
flaky: if main looper is not preferred, test fails. So if we run
all Telephony unittests, it happens to pass; but if we run
MultiSimSettingControllerTest separately, it fails.

Bug: 132909128
Bug: 130687339
Test: unittest
Change-Id: I9bcd4424f9999f1abc9711c41ce1d71f7ca0a57b
Merged-In: I9bcd4424f9999f1abc9711c41ce1d71f7ca0a57b