Allow mobile data on for grouped primary subscriptions.
We have two rules now:

1) If multiple non-oppt subscriptions are active, disable mobile data
for those other than default data sub.
2) When mobile data setting changes in a subscription, the corresponding
value of other subscriptions in the same group are also changed.

So if there are two grouped primary subscriptions, we end up disabling
mobile data on both. Which is bad. So we are making the exception to
allow mobile data to be on both if they are grouped.

Bug: 132204817
Test: unittest
Change-Id: Ie01bcb948d2f63e694cf99f36a7cb2edb0fc0acb
Merged-In: Ie01bcb948d2f63e694cf99f36a7cb2edb0fc0acb