Improve CDNR log and remove duplicated codes
Add some logs to the legacy CDNR, also add the spn update records to the
local log so we will not lose the records due to the limited logcat
buffer size.

This also removed the duplicated codes from the legacy CDNR. Both the new
CDNR and legacy CDNR will use the same method to send the carrier
display name update.

Bug: 122921456
Test: build
Change-Id: I8ecf41ed7911a68ffef3d0e0b281a1e93ae0259e