Remove UID packageName check in checkCarrierPrivilegeOnSubList.
When groupUuid is updated through carrier config update, it calls
checkCarrierPrivilegeOnSubList but not through a Binder call. So
UID and packageName check will fail. In this case, the check is not
needed as we get packageName through binding carrier service so it's
already a trusted source.
Here we are:
1) removing the check in the checkCarrierPrivilegeOnSubList.
2) make sure all external calls checks UID / packageName before calling
checkCarrierPrivilegeOnSubList.

Bug: 132968722
Test: manual
Change-Id: Ibced4799cfa4d721e17ce69aa87f884613a5dc26