Fix bug of disabling grouped CBRS during profile switch on primary SIM
When we update subInfo, we remove the info first and re-fill the fields. In between, we
unexpectedly believed it's deactivated which results in deactivating its grouped oppt sub
and broadcast the decision out. Two actions:
1) Make disableGroupedOppt subscription handled asynchronous in MultiSimSettingController
after all sub info changes are done.
2) Don't broadcast oppt subInfo change whenever the cache is updated, because it may update
several times in one operation. Instead, wait until notifySubscriptionInfoChanged is
triggered, and only broadcast if oppt subInfo is changed.

Bug: 131919246
Test: manual
Change-Id: Ie0b6d84d12c7fa3ede6e8b24275145849a89f0a3