Accept restricted network request on non default data SIM.
This allows system apps to setup data connection on non default data
SIM temporarily, for usercases such as VVM.

We used to block those requests in PhoneSwitcher side. By removing it,
network request will be sent to DcTracker, which will only accept
default network request if user data is on or if it's restricted request.
Which serves our purpose.

Bug: 132450451
Test: unittest
Change-Id: I361f11d5a0975e93ef1beb9f6780bb13b22e5cd7