Suppress SIM app dialog for locked devices and with bundled apps.
The platform supports a mechanism to bundle carrier apps but keep them
fully uninstalled until a matching SIM is inserted (see
CarrierAppUtils). However, the installation of such apps races against
the SIM app dialog which prompts the user to install that app, even as
it's being actively installed from the bundled APK.

To avoid this race, the SIM app dialog should treat
hidden-until-installed components the same way as it treats installed
components and avoid showing a prompt. The assumption is that if the
OEM configured an app in this state, then some external process should
be responsible for enabling it, and we don't need to prompt the user
about it.

Furthermore, if the user has yet to unlock credential-encrypted
storage, we cannot prompt them to install an app as the app store is
unlikely to be direct-boot-aware. We treat this case similarly to when
Setup Wizard hasn't run yet - wait for the user to unlock the device,
and then show a notification instead of a full-screen activity.

Bug: 134067502
Bug: 134067994
Test: Validated both scenarios on device
Change-Id: Ifd2df776982df6d83b316990c97d90b6c980d43d