Make sure on non-dds phone, try MMS data connection when PS is not attached.
1) mAutoAttachEnabled is not updated for certain scenarios. So instead
of updating it true for non-dds case, we always check before setting up
data connection, that if it's on non-dds and voice is registered on
non-LTE network, if true we'll setup the data connection even if it's
not attached.

2) If data is not registered, we use voice rat to find the matching APN.
The criteria is APN supports a rat that maps to the same accessNetworkType
as voice rat.

Bug: 134086103
Test: manual
Change-Id: If3c80d5745b956f1c2b1b79d480bf2c2680d8923
Merged-In: If3c80d5745b956f1c2b1b79d480bf2c2680d8923