[Encrypted IMSI] Make WLAN the default certificate type
The 3GPP spec for encrypted IMSI certificate specifies
"key-identifier" and "public-key" String properties, but does not
specify "key-type" property which was a mandatory requirement in the
implementation prior to this change. Make the "WLAN" key type as the
default type, so if it is not provided by the carrier, then the logic
would treat the certificate as WLAN type.

Bug: 134604306
Test: Install carrier SIM and auto connect to carrier Wi-Fi network
Test: atest CarrierKeyDownloadMgrTest
Change-Id: Ic749eb7ca956ff7ef6d86e481f56f95562c5dced