Fix issue with syncing default voice sub with Telecom.
When setting the default voice sub ID we would not try to set the phone
account in Telecom if the subid setting didn't change.  There are some
instances we've seen in the field where the sub ID is set (but not changed)
and Telecom needs to update the default outgoing phone account.

Since the code alread checks to see if the phone account did change, we
remove the check for setting equality and always try to update telecom
here intead.

Test: Switch between single-sim and DSDS mode; confirm sync is not lost.
Bug: 134469771
Change-Id: Ieb62c46a8fd6b09eb4778ebadd81492a9c6dd234