Update isEmptyProfile of UiccProfile.
If no UiccApplications is reported from modem, we assert it's empty
profile. But at start of SIM_REFRESH we internally clear UiccApplications
which sometimes confuses us to believe it's empty profile. So we'll add
a cache of last reported num of UiccApplications from modem, which is
more reliable to determine whether modem has landed on empty profile or
not.

Bug: 131919246
Test: manual
Change-Id: I9a3fccbd4af9dfcfdb6ede04e6c1926a0d392e9a