Allow data switch in emergency call.
Remove the restriction that no data switching can happen in emergency
call at all.
In the scenario of DSDS, if the mergency call is placed on non-dds
subscription, we could setup data conneciton on it if data is allowed.
So we should still re-evaluate data sub even in emergency call. The
only thing we need to make sure is, in older devices where on-demand
ps attach / detach is needed, don't detach the phone in voice call.

Bug: 134073343
Test: sanity
Change-Id: Ie2549f2cd426c902935a07afd90a4771b8498f8f
Merged-In: Ie2549f2cd426c902935a07afd90a4771b8498f8f