Reduce log spam in DcTracker and DataEnabledSettings
For a device in a normal state (connected), just opening Settings >
Network & internet triggers this log, which can flush useful info out of
the local log buffer. We can keep the CONNECTING and DISCONNECTING logs
because those are less common but we should remove the log for
CONNECTED.

In DataEnabledSettings we also only local log when the setting actually
changes.

Bug: 133852635
Test: manual
Change-Id: I3a231db8428c26263f892aa5ea0ebe89222cdd61