Fix that PhoneSwitcher only listens to call state change of default sub.
PhoneSwitcher currently listens to precise calls state change through
PhoneStateListener, which without specifying a subId will monitor call
state of default voice subscription. This is enough for CBRS data switching
as default voice subscription is always the primary, but not enough as
we now support switching in DSDS case.
So here we are making the change to listen to precise call state change
on both phones (CS or PS call).

Bug: 134981077
Test: manual and unittest
Change-Id: Idd21283ba3f01bc99fd9afff7c83a79711649f01
Merged-In: Idd21283ba3f01bc99fd9afff7c83a79711649f01