Make NewNitzStateMachine the only impl
Make NewNitzStateMachine the only impl by deleting
OldNitzStateMachine and associated classes and renaming
'New' classes. The "Old" code was kept around in case
the changes in the "New" code turned out to cause problems
in the Q cycle. Now we can embed the new behavior and
remove the maintenance overhead of having two classes.

Bug: 133492648
Test: build only
Change-Id: I81ad0c324dab64913b943a20b3e9e30e344b0c92