Switch to SIM with voice call on hold.
In DSDS mode, when a SIM is in voice call, we might switch data
to it if user wants to. There's an issue that we don't switch
if user places the call on hold, in which case the user actually
loses data connections. Here the fix is we also switch data if
there's a call on hold.

Bug: 136843023
Test: manual and unittest
Change-Id: I43e3423748dd6492c4961b08b4d8c4f99a9d75a4
Merged-In: I43e3423748dd6492c4961b08b4d8c4f99a9d75a4