Always send USSD codes over the CS pipe.
A change was made early in the Q cycle which would try to send USSD codes
over the IMS pipe when CS was not in service.  It was not possible to test
that change at the time.  However, it turns out there are limitations in
how the modem handles the requests and it should just be sent over CS
all the time.

Test: Manual test using carrier network.
Bug: 135194912
Change-Id: I8a2b4bae812aa19562fcffe5211edda35648d38e