Extract PLMN from Voice CellIdentity for Locale
The locale tracker needs to know what MCCs are seen,
and it doesn't matter whether the device registers
successfully on a network, simply camping on a tower
(even for emergency services only) is a strong indicator
of the current country. This change tries to use a
registered network operator from the IRadio::getOperator
command, but if that fails, it falls back to using the
PLMN from the voice operator which is reported even when
the device only camps for emergency services.

Bug: 136036974
Test: atest LocaleTrackerTest
Merged-In: I634aeb9fb95cc361b21b79f209a90dddf18f5896
Change-Id: I634aeb9fb95cc361b21b79f209a90dddf18f5896
(cherry picked from commit 09ee2e53c9259fd4d0afd8b6357409744e146e98)