De-bounce Operator Numeric From ServiceState
ServiceState changes can happen rapidly and cause
transient cases where the operator numeric becomes
invalid, but for brief service interruptions, there
is no need to assume that the locale has changed.

Furthermore, in case the modem is searching for service,
it may provide PLMN info briefly when it camps but then
lose that PLMN as it continues to search. Again, this
causes significant noise in the signal coming from
ServiceStateTracker that doesn't actually imply that
the phone has moved to a new country. To avoid constantly
going in-and-out of service based on SST changes, add
a 10 minute delay to clearing the info from SST.

Bug: 136036974
Test: atest LocaleTrackerTest
Change-Id: Icada0cda8e947460d9a4ca3bc18c6acfd62548d6