Notify service state change when subId goes valid
Since we no longer broadcast service state changes for invalid subIds,
there's a possibility that a caller may miss service state indications
if the service state is broadcast before the subId is set (this is a
race condition). We can address this by rebroadcasting the service state
in the case when the subId goes from invalid to valid.

Bug: 133140128
Test: Checked that service state is rebroadcast when subid becomes
    valid, b/138357571 did not show up 5/5 times
Change-Id: I42ee6940f7d5b8458e8a17cde0a88a4cf156a092