Move isDefaultOkForCountryTimeZoneDetection
Push the CountryTimeZones.isDefaultOkForCountryTimeZoneDetection(long)
method up to telephony code. When written it was unclear whether the
method was going to be useful for future time zone detection code but
it was "cheap" to put it in the libcore code / Core Platform API. It is
still unclear whether it will be used but it's far from certain and
every API method will carry a support cost in future so pushing it up
to the one user and making the libcore code more general makes more
sense.

This commit also contains an improvement to
CountryTimeZones.getTimeZone() logic and makes it part of the
Core Platform API.

Bug: 139928367
Test: atest com.android.internal.telephony.TimeZoneLookupHelperTest
Change-Id: I954a42208a013fd009f1ad6541b6c98d7cc94f6f