Track changes in libcore's API
Track changes in libcore's API lookupByOffsetWithBias().

The method args have been configured so that the behavior should
stay the same for most cases. Only if the NITZ DST information is
missing will there be a change, and it should mean better matching
behavior. See the comment added in the code for more information.

Bug: 139091367
Bug: 140712361
Test: treehugger
Change-Id: I0b9ff076e1d60196477fab404578100eb89c548f