Additional SGP.22 v2.1+ BPP validity checking.
As of SGP.22 version 2.1 and newer, prior to sending any APDU to the
eSIM, the bound profile package should be validated with additional
criteria. Mainly, it should not contain any extraneous ASN.1 tags or
have zero profile elements.

SGP.22 version 2.0 does not mention what should happen in this case, so
we leave that behavior unmodified.

Added unit tests to EuiccCardTest and verified with Comprion (SGP.23
testing tool).

Test: atest EuiccCardTest
Test: SGP.23 4.4.25.2.1_03 and 4.4.25.2.1_05
Bug: 140314189
Change-Id: I6ebe6d64496cc62d86652a833d04805bff202862